kernel.mk: fix bison unable to find m4 file issue
Android builds a version of bison which is unusable outside the tree.
Since the bison is invoked in the KBUILD_OUTPUT dir, the error is
generated:

bison: external/bison/data/m4sugar/m4sugar.m4: cannot open: No such file or directory

To workaround it, add a symlink to the external dir so that bison can
get the expected m4 file.