init.sh: allow HAS_SENSORS to override all cases
HAS_SENSORS in cmdline was introduced in commit f356d114 to override
the default has_sensors variable. However, it only take effects if
the sensors type is not kbd. This is not my original idea.

The change allows HAS_SENSORS to override the default value no matter
what the sensors type is.