init.sh: enable virt_wifi if needed
Some Android apps are not interested in ethernet devices and would rather
connect to a wireless interface. This patch allows the devices with only
ethernet interface to treat one of its network connections as wireless,
satisfying Android apps's network connection requirements.

The ethernet interface being simulated can be specified by the property
net.virt_wifi. The default is eth0.

The feature is enabled if wlan0 doesn't exist, or VIRT_WIFI=1 is set in
kernel cmdline. To disable it, set VIRT_WIFI=0 kernel cmdline.