HACK: granting full permission for apps data (1/2)
This is a temporary hack to combat against the apps permission issue on Android 11. Whenever an app is open it will always enable a property (persist.sys.nodataperm) to make it execute a script to grant the data full permissions (777).  Keep in my that this is a temporary hack and there should be a better solution like using bind-mounting filesystem.

Thanks PrimeOS team for this solution