Use copy-firmware.sh script to copy linux firmwares
Instead of the old method which use PRODUCT_COPY_FILES, we use the
script that is already inside linux-firmware called copy-firmware.sh
to copy firmwares instead. This script will read a file called WHENCE
and not only copy the firmware but also symlink them so that drivers
can correctly recognize it.