Introducing a way to enable hidden features on PC
When these option set to 1, it will turn on some features on AOSP
that is made specifically for PCs. For example :
- If PC_MODE=1 is set, it will set the hardware as PC with lots of
functions that is tailor made for PCs like freeform windows.
- If HPE=1 is set, it will be recognize as Google Play Games for PC.
Stores like Play Store or even Aurora Store will only show games that
is on GPG for PC and nothing else. This is useful for people who want
to get native version of some games like Arknights.

In order to activate the feature, we bind mount the xml
into /system/etc/permissions or /system/etc/sysconfig