Building Celadon's new sensors HIDL
This is a new HIDL created by Intel, more specifically Project Celdon
team. Originally it was designed to use libiio's network backend on a VM
to communicate with the host. But we take it and use the local backend
of libiio instead.

We also switch android.hardware.sensors to hwbinder and keep sensors@1.0
because we also want to switch between the old HALs and the new HIDL

WARNING : due to the fact that kbd is having a problem making the whole
OS bootloop, DO NOT MERGE THIS COMMIT unless you use it for testing purpose

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>