Refactoring init_hal_media
A major part that I refactored is that removal or FFMPEG_CODEC,
instead of using it to turn on both OMX and codec2 of the ffmpeg
codecs, I separate them into FFMPEG_OMX_CODEC and FFMPEG_CODEC2_PREFER.
FFMPEG codec2 is already automatically enabled, just need a line to make
it as the default.
The rest we have these new props :
- OMX_NO_YUV420 : it just NO_YUV420 renamed.
- FFMPEG_CODEC2_DEINTERLACE & FFMPEG_CODEC2_DEINTERLACE_VAAPI : change the
deinterlace mode for ffmpeg codecs in both software & hardware decoder
via FFMPEG.
- FFMPEG_CODEC2_DRM : turn on/off DRM prime handle on ffmpeg codecs, 0 is
the default because not any Gralloc supporting it.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>