Temporary putting a script to modprobe at early-init stage
Due to the fact that we are forced to use updatable APEX, uevent modprobe
started to become unreliable as libc may not be prepared yet. Temporary
make a modprobe script using toybox modprobe and run it at early-init.

This is not the definitive solution, we need to find a better way.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>